extract terminal.dart
